feat: grid_edge_point_scalars added and tested

+++ src/cglib/src/scalar.h
@@ -1,9 +1,8 @@
 #include <Eigen/Dense>
 #include "grid/cell.h"
 
-// Declare the grid_edge_point_scalars function
-Eigen::ArrayXf grid_edge_point_scalars(
+Eigen::ArrayXd grid_edge_point_scalars(
     const Eigen::ArrayXi& edge_ndindex,
     int edge_axis,
-    const Eigen::ArrayXf& grid_scalars_flattened,
+    const Eigen::ArrayXd& grid_scalars_flattened,
     const Eigen::ArrayXi& grid_cell_ndcount);
